Rename "repertory" module in "inventory" and refactor Sanitize function
