+++ cli/cli.go
@@ -39,7 +39,7 @@ import (
 	"github.com/arduino/arduino-cli/cli/upload"
 	"github.com/arduino/arduino-cli/cli/version"
 	"github.com/arduino/arduino-cli/configuration"
-	"github.com/arduino/arduino-cli/repertory"
+	"github.com/arduino/arduino-cli/inventory"
 	"github.com/mattn/go-colorable"
 	"github.com/rifflock/lfshook"
 	"github.com/sirupsen/logrus"
@@ -168,8 +168,8 @@ func preRun(cmd *cobra.Command, args []string) {
 	configuration.Init(configPath)
 	configFile := viper.ConfigFileUsed()
 
-	// initialize repertory
-	repertory.Init(viper.GetString("directories.Data"))
+	// initialize inventory
+	inventory.Init(viper.GetString("directories.Data"))
 
 	//
 	// Prepare logging
