+++ commands/compile/compile.go
@@ -47,7 +47,7 @@ func Compile(ctx context.Context, req *rpc.CompileReq, outStream, errStream io.W
 
 	tags := map[string]string{
 		"fqbn":            req.Fqbn,
-		"sketchPath":      telemetry.SanitizeSketchPath(req.SketchPath),
+		"sketchPath":      telemetry.Sanitize(req.SketchPath),
 		"showProperties":  strconv.FormatBool(req.ShowProperties),
 		"preprocess":      strconv.FormatBool(req.Preprocess),
 		"buildProperties": strings.Join(req.BuildProperties, ","),
